like btn

+++ app/Repositories/UserApi/PostRepository.php
@@ -1,7 +1,7 @@
 <?php
 
 namespace App\Repositories\UserApi;
-
+use Illuminate\Http\Request;
 use App\Like;
 use App\Post;
 use App\Friendship;
@@ -21,68 +21,35 @@ class PostRepository {
         $this->request = $request;
     }
 
-
-    public function create() 
-    {
-        $new_status = $this->post->create([
-            'user_id' => auth()->user()->id,
-            'caption' => $this->request->input('caption')
-        ]);``
-
-        return "Sukses Menambahkan Status";
-    }
-
-    public function delete() 
-    {
-        $delete = $this->post->find($this->request->input('post_id'));
-        $delete->delete();
-
-        return "Sukses menghapus post";
-    }
-
-    public function list() 
-    {
-        $friend = $this->friend
-                        ->where('user_id', auth()->user()->id)
-                        ->get(['friend_id']);
-
-        $result = $this->post
-                        ->whereIn('user_id', $friend)
-                        =>get();
-
-        $response = [];
-        $response['data'] = $result->map(function($post){
-            'post_id' => $post->id,
-            'username' => $post->user->username,
-            'full_name' => $post->user->fullname,
-            'caption' => $post->caption,
-            'comment_count' => $post->countComment(),
-            'recent_comment' => $post->recentComment()
-        });
-
-        return $response;
-    }
-
     public function like()
     {
         $post_id = $this->request->input('post_id');
+
+        $post = Post::find($post_id);
+        if($post->isLiked()) return ['error' => true, 'message' => 'anda sudah menambahkan like'];
+
         $newLike = $this->like->create([
             'post_id' => $post_id,
             'user_id' => auth()->user()->id
         ]);
         
         $post = $this->post->find($post_id);
-        return ['error' => false, 'message' => 'sukses menambahkan like', 'count_like' => $post->like_count()];
+        $count = ($post->likeCount() <= 1) ? ' Star' : ' Stars';
+        return ['error' => false, 'message' => 'sukses menambahkan like', 'count_like' => $post->likeCount().$count];
     }
 
     public function unlike()
     {
         $post_id = $this->request->input('post_id');
 
+        $post = Post::find($post_id);
+        if(!$post->isLiked()) return ['error' => true, 'message' => 'anda sudah menghapus like'];
+
         $liked = $this->like->where('user_id', auth()->user()->id)->where('post_id', $post_id)->firstOrFail();
         $liked->delete();
 
         $post = $this->post->find($post_id);
-        return ['error' => false, 'message' => 'sukses menghapus like', 'count_like' => $post->like_count()];
+        $count = ($post->likeCount() <= 1) ? ' Star' : ' Stars';
+        return ['error' => false, 'message' => 'sukses menghapus like', 'count_like' => $post->likeCount().$count];
     }
 }
